% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{author_fix}
\alias{author_fix}
\title{Manual author name fixes for \code{pkg_db()}}
\usage{
author_fix()
}
\description{
The output is a named character vector where the name is the exact match with the
"Author" column and the value is what it should be replaced as.
}
\seealso{
pkg_db_clean_author, pkg_db_manual_author_check
}
