% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranscrub.R
\name{cranscrub}
\alias{cranscrub}
\title{Package downloads from the RStudio CRAN mirror with duplicates removed}
\usage{
cranscrub(
  packages = NULL,
  when = c("last-day", "last-week", "last-month"),
  from = "last-day",
  to = "last-day"
)
}
\arguments{
\item{packages}{A character vector of CRAN package names}

\item{when}{Either \code{last-day}, \code{last-week} or \code{last-month}. If this is
given then \code{from} and \code{to} are ignored. NOT implemented yet!}

\item{from}{Start date in YYYY-MM-DD format.}

\item{to}{End date in YYYY-MM-DD format.}
}
\description{
The arguments behave the same as \code{cranlogs::cran_downloads}.
}
\examples{
cranscrub(packages = "fable", from = "2020-01-03", to = "2020-01-05")
}
