% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskview.R
\name{ctv_names}
\alias{ctv_names}
\title{Get all the list of topics in the CRAN task views}
\source{
https://cran.r-project.org/web/views/
}
\usage{
ctv_names(show_topic = FALSE)
}
\arguments{
\item{show_topic}{A logical value to indicate whether to return the topic or name.
Default is FALSE and returns the task view name.}
}
\description{
Get all the list of topics in the CRAN task views
}
