% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskview.R
\name{ctv_pkgs}
\alias{ctv_pkgs}
\title{Get the list of packages for a given task}
\usage{
ctv_pkgs(topic = ctv_topics(), repos = NULL)
}
\arguments{
\item{topic}{A topic listed in the CRAN task view.}

\item{repos}{A base URL of the repository. By default, \code{getOption("repos")} is tried
and otherwise \code{getOption("CRAN")} is used.}
}
\description{
Get the list of packages for a given task
}
