% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskview.R
\name{ctv_table}
\alias{ctv_table}
\title{Get the CRAN task view as a table}
\usage{
ctv_table(repos = NULL)
}
\arguments{
\item{repos}{A base URL of the repository. By default, \code{getOption("repos")} is tried
and otherwise \code{getOption("CRAN")} is used.}
}
\description{
Get the CRAN task view as a table
}
