% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskview.R
\name{ctv_topics}
\alias{ctv_topics}
\title{Get all the list of topics in the CRAN task views}
\source{
https://cran.r-project.org/web/views/
}
\usage{
ctv_topics()
}
\description{
Get all the list of topics in the CRAN task views
}
