% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ctvExperimentalDesign}
\alias{ctvExperimentalDesign}
\title{The CRAN logs for the RStudio mirror for the CRAN task view Experimental Design}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
\item{date}{date}
\item{package}{the package name}
\item{n_unique}{the number of unique downloads}
\item{n_total}{the number of total downloads. This should be the same as \code{cranlogs::cran_downloads()}.}
}
}
\usage{
ctvExperimentalDesign
}
\description{
The CRAN logs for the RStudio mirror for the CRAN task view Experimental Design
}
\examples{
library(ggplot2)
ggplot(ctvExperimentalDesign, aes(n_unique, n_total)) + 
  geom_point() + 
  geom_abline(intercept = 0, slope = 1, color = "red")
}
\keyword{datasets}
