% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{pkg_db}
\alias{pkg_db}
\title{Get the package database}
\usage{
pkg_db(pkgs)
}
\arguments{
\item{pkgs}{A character vector of the name of CRAN packages.}
}
\description{
All CRAN mirrors contain the packages.rds file and this downloads the one
in the RStudio CRAN mirror.
}
