% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{pkg_db_clean_author}
\alias{pkg_db_clean_author}
\title{Clean the Author names from \code{pkg_db()}}
\usage{
pkg_db_clean_author(db, replace = author_fix())
}
\arguments{
\item{db}{The data from \code{pkg_db()} output}

\item{replace}{A named character vector to replace the Author value with. The ones
that were manually fixed are contains in \code{author_fix()}.}
}
\description{
Clean the Author names from \code{pkg_db()}
}
