% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{pkg_db_clean_imports}
\alias{pkg_db_clean_imports}
\title{Clean the "Imports" column in \code{pkg_db()}}
\usage{
pkg_db_clean_imports(db)
}
\arguments{
\item{db}{The data from \code{pkg_db()} output}
}
\description{
Clean the "Imports" column in \code{pkg_db()}
}
