% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{pkg_db_clean_suggests}
\alias{pkg_db_clean_suggests}
\title{Clean the "Suggests" column in \code{pkg_db()}}
\usage{
pkg_db_clean_suggests(db)
}
\arguments{
\item{db}{The data from \code{pkg_db()} output}
}
\description{
Clean the "Suggests" column in \code{pkg_db()}
}
