% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{pkg_db_manual_author_check}
\alias{pkg_db_manual_author_check}
\title{Manually check Author entry from \code{pkg_db()}}
\usage{
pkg_db_manual_author_check(db, replace = author_fix(), start = 1)
}
\arguments{
\item{db}{The data from \code{pkg_db()} output}

\item{replace}{A named character vector to replace the Author value with. The ones
that were manually fixed are contains in \code{author_fix()}.}

\item{start}{The starting row index corresponding to \code{db}.}
}
\description{
There are so many issues with the author entry in database from \code{pkg_db()}
that manual checks are unavoidable. This function prints out the original
Author entry and contrast to that of replaced one, \code{Author@R} and the cleaned up
version.
}
