% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranlog.R
\name{read_cran_log}
\alias{read_cran_log}
\title{Get the unprocessed RStudio CRAN log}
\usage{
read_cran_log(date)
}
\arguments{
\item{date}{The date given either as a string format of "YYYY-MM-DD" or
a date object.}
}
\description{
Get the unprocessed RStudio CRAN log
}
